[Bug] - Transactions were being created with Tag ids
Our query strategy here is to join tags and transactions. First, transactions.id is renamed to transactions.transaction_id so we can do a natural join. This means that when we construct a Transaction type from this result relation, the id attribute is actually a Tag id.

This can be avoided by first renaming tags.id to tags.tag_id, joining, and then renaming transactions.transaction_id to transactions.id on the result. A little tricky, but this ensures that Transactions are created with their proper ids.

Reproduced this bug with a property-based test that strictly uses the Application API to create and query data (one exception was calling a use case method directly, since the Application API is still under a bit of construction). Instead of generating complex data, we generate sequences of API calls and perform them. This feels like a good pattern for the future.