Abstract sequences of application actions.
This allows tests to simply specify which actions they want to perform, and what should always be true for all sequences of them. Note, hypothesis is still required for the check! block since there's often still random values to be read after the state has been created. It would be nice to avoid this, but I don't know the best way to do that for now.